requirements.txt
fastapi[standard]
sqlalchemy
psycopg2-binary
pydantic
python-jose
passlib
python-multipart
slugify-python
dotenv
alembic
